[REF] separate authen & password in account
