+++ apps/account/src/controllers/account.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AccountController } from './account.controller';
-import { AccountService } from './account.service';
+import { AccountService } from '../services/account.service';
 
 describe('AccountController', () => {
   let accountController: AccountController;
