+++ apps/account/src/services/account.service.ts
@@ -1,18 +1,18 @@
 import { BadRequestException, Inject, Injectable } from '@nestjs/common';
 import * as bcrypt from 'bcrypt';
-import { RegisterAccountDto } from "./dto/register.account.dto";
+import { RegisterAccountDto } from "../dto/register.account.dto";
 import { IServiceResponse } from "@app/rabbit";
-import { AccountEntity } from './entity/account.entity';
+import { AccountEntity } from '../entity/account.entity';
 import { Repository, UpdateResult } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Database } from '@app/database';
-import { ACCOUNT_MESSAGE_DB_RESPONSE } from './constant/account-patterns.constants';
-import { LoginAccountDto } from './dto/login.account.dto';
+import { ACCOUNT_MESSAGE_DB_RESPONSE } from '../constant/account-patterns.constants';
+import { LoginAccountDto } from '../dto/login.account.dto';
 import { CognitoService } from '@libs/cognito';
-import { AuthChangePasswordUserDto } from './dto/change-password.account.dto';
-import { TokenAccountDto } from './dto/token.account.dto';
-import { AuthVerifyUserDto } from './dto/verify-email.account.dto';
-import { RefreshTokenAccountDto } from './dto/refresh-token.account.dto';
+import { TokenAccountDto } from '../dto/token.account.dto';
+import { AuthVerifyUserDto } from '../dto/verify-email.account.dto';
+import { RefreshTokenAccountDto } from '../dto/refresh-token.account.dto';
+import { MemberEntity } from 'apps/member/src/entity/member.entity';
 
 @Injectable()
 export class AccountService {
@@ -21,7 +21,7 @@ export class AccountService {
 
   constructor(
     @InjectRepository(AccountEntity, Database.PRIMARY) private accountRepository: Repository<AccountEntity>,
-    @Inject(CognitoService) private awsCognitoService: CognitoService) {}
+    @Inject(CognitoService) private awsCognitoService: CognitoService) { }
 
   async findByEmail(email: string): Promise<IServiceResponse<AccountEntity>> {
     const account = await this.accountRepository.findOneBy(
@@ -37,34 +37,43 @@ export class AccountService {
     }
   }
 
-  async register(createAccoutDto: RegisterAccountDto): Promise<IServiceResponse<AccountEntity>> {
+  async register(createAccoutDto: RegisterAccountDto, memberEntity: MemberEntity): Promise<IServiceResponse<AccountEntity>> {
+    try {
 
-    const accountExist = await this.findByEmail(createAccoutDto.email);
-    const password = createAccoutDto.password
+      const accountExist = await this.findByEmail(createAccoutDto.email);
+      const password = createAccoutDto.password
 
-    if (accountExist.state) {
+      if (accountExist.state) {
+        return {
+          state: accountExist.state,
+          data: accountExist.data,
+          message: ACCOUNT_MESSAGE_DB_RESPONSE.EXISTING_EMAIL
+        };
+      }
+
+      const hash = await bcrypt.hash(createAccoutDto.password, this.saltOrRounds);
+      createAccoutDto.password = hash
+      createAccoutDto.member_id = memberEntity
+      const accountEntity = this.accountRepository.create(createAccoutDto);
+      const result = await this.accountRepository.save(accountEntity);
+
+      await this.awsCognitoService.registerUser(
+        createAccoutDto.email,
+        password
+      )
+
       return {
-        state: accountExist.state,
-        data: accountExist.data,
-        message: ACCOUNT_MESSAGE_DB_RESPONSE.EMAIL_FOUND
+        state: !!result,
+        data: result,
+        message: ACCOUNT_MESSAGE_DB_RESPONSE.CREATED
+      };
+
+    } catch (e) {
+      return {
+        state: false,
+        data: e.name
       };
     }
-
-    const hash = await bcrypt.hash(createAccoutDto.password, this.saltOrRounds);
-    createAccoutDto.password = hash
-    const accountEntity = this.accountRepository.create(createAccoutDto);
-    const result = await this.accountRepository.save(accountEntity);
-
-    await this.awsCognitoService.registerUser(
-      createAccoutDto.email,
-      password
-    )
-
-    return {
-      state: !!result,
-      data: result,
-      message: ACCOUNT_MESSAGE_DB_RESPONSE.CREATED
-    };
   }
 
   async login(loginAccountDto: LoginAccountDto): Promise<IServiceResponse<TokenAccountDto>> {
@@ -104,7 +113,7 @@ export class AccountService {
         authVerifyUserDto.confirmationCode,
       ).then(
         result => {
-            updateResult = this.accountRepository.update(
+          updateResult = this.accountRepository.update(
             {
               email: authVerifyUserDto.email
             },
@@ -166,83 +175,5 @@ export class AccountService {
     }
 
   }
-
-  // TODO MOVE THIS INTO A SERVICE
-  async changePassword(authChangePasswordUserDto: AuthChangePasswordUserDto): Promise<IServiceResponse<UpdateResult>> {
-
-    try {
-
-      await this.awsCognitoService.changeUserPassword(
-        authChangePasswordUserDto.email,
-        authChangePasswordUserDto.currentPassword,
-        authChangePasswordUserDto.newPassword,
-      )
-
-      let result = null
-      result = this.updatePassword(
-        authChangePasswordUserDto.email,
-        authChangePasswordUserDto.newPassword
-      ).then(
-        updateRsult => { result = updateRsult }
-      )
-
-      return {
-        state: true,
-        data: result
-      }
-    } catch (e) {
-      return {
-        state: false,
-        data: e.name,
-        message: ACCOUNT_MESSAGE_DB_RESPONSE.NOT_FOUND
-      };
-    }
-
-  }
-
-  async resetPassword(authConfirmPasswordUserDto: AuthChangePasswordUserDto): Promise<IServiceResponse<UpdateResult>> {
-
-    try {
-
-
-      await this.awsCognitoService.confirmUserPassword(
-        authConfirmPasswordUserDto.email,
-        authConfirmPasswordUserDto.confirmationCode,
-        authConfirmPasswordUserDto.newPassword,
-      )
-
-      let result = null
-      this.updatePassword(
-        authConfirmPasswordUserDto.email,
-        authConfirmPasswordUserDto.newPassword
-      ).then(
-        updateRsult => { result = updateRsult }
-      )
-
-      return {
-        state: true,
-        data: result
-      }
-    } catch (e) {
-      return {
-        state: true,
-        data: null,
-        message: ACCOUNT_MESSAGE_DB_RESPONSE.NOT_FOUND
-      };
-    }
-
-  }
-
-  async updatePassword(email: string, password: string): Promise<UpdateResult> {
-    return await this.accountRepository.update(
-      {
-        email: email
-      },
-      {
-        password: await bcrypt.hash(password, this.saltOrRounds)
-      }
-    );
-  }
-
 
 }
